+++ R/z_print_auto.R
@@ -1,25 +1,14 @@
-.as_object_auto = function(df1, destination = NULL, ...) {
+.print_auto = function(df1, destination = NULL, ...) {
   dest = .get_destination(destination = destination)
   assert_that(dest %in% c("", "html", "latex"))
 
   if (dest == "") {
-    .as_object_huxtable(df1, ...)
+    .print_huxtable(df1, destination = dest, ...)
   } else if (dest == "html") {
-    .as_object_gt(df1, ...)
+    .print_gt(df1, destination = dest, ...)
   } else if (dest == "latex") {
-    .as_object_kableextra(df1, destination = "latex", ...)
-  }
-}
-
-.print_auto = function(hh, destination = NULL, ...) {
-  dest = .get_destination(destination = destination)
-  assert_that(dest %in% c("", "html", "latex"))
-
-  if (dest == "") {
-    .print_huxtable(hh, destination = "", ...)
-  } else if (dest == "html") {
-    .print_gt(hh, destination = "html", ...)
-  } else if (dest == "latex") {
-    .print_kableextra(hh, ...)
+    .print_kableextra(df1, destination = dest, ...)
+  } else {
+    stop("?")
   }
 }