+++ R/z_print_huxtable.R
@@ -1,6 +1,23 @@
-.as_object_huxtable = function(df1, ...) {
-  assert_package("as_object", "huxtable")
+.print_huxtable = function(df1, destination = NULL, ...) {
+  ##
+  if (inherits(df1, "surveytable_list")) {
+    if (length(df1) > 0) {
+      for (ii in 1:length(df1)) {
+        Recall(df1 = df1[[ii]], destination = destination, ...)
+      }
+    }
+    return(invisible(NULL))
+  }
 
+  ##
+  assert_package("print", "huxtable")
+  assert_that(inherits(df1, "surveytable_table"))
+  dest = .get_destination(destination = destination)
+  assert_that(dest != "latex"
+              , msg = "Have not implemented LaTeX printing with huxtable yet. Try set_opts(output = 'kableExtra')")
+  assert_that(dest %in% c("", "html"))
+
+  ##
   hh = huxtable::set_all_borders( huxtable::hux(df1) )
   if (!is.null(txt <- attr(df1, "title"))) {
     huxtable::caption(hh) = txt
@@ -11,16 +28,8 @@
   if (!is.null(txt <- attr(df1, "footer"))) {
     hh %<>% huxtable::add_footnote(txt)
   }
-  hh
-}
 
-.print_huxtable = function(hh, destination = NULL, ...) {
-  assert_package("print", "huxtable")
-  dest = .get_destination(destination = destination)
-  assert_that(dest != "latex"
-              , msg = "Have not implemented LaTeX printing with huxtable yet. Try set_opts(output = 'kableExtra')")
-  assert_that(dest %in% c("", "html"))
-
+  ##
   if (dest == "") {
     gow = getOption("width")
     op_ = options(width = 10)
@@ -28,8 +37,10 @@
     huxtable::print_screen(hh, colnames = FALSE, min_width = 0
         , max_width = max(gow * 1.5, 150, na.rm = TRUE))
     cat("\n")
-  } else {
+  } else if (dest == "html") {
     huxtable::print_html(hh)
+  } else {
+    stop("?")
   }
 }