+++ R/zzz.R
@@ -24,8 +24,7 @@ env = new.env()
     , surveytable.drop_na = FALSE
     , surveytable.max_levels = 20
     , surveytable.csv = ""
-    , surveytable.output_object = ".as_object_auto"
-    , surveytable.output_print = ".print_auto"
+    , surveytable.print = ".print_auto"
 
     ## other
     , surveytable.tx_prct = ".tx_prct"