+++ NEWS.md
@@ -6,6 +6,8 @@
 * Conditional independence test: `tab_subset()` with argument `test` set to the 
 value of interest. 
 * Confidence intervals for numeric variables. 
+* Excel tables and charts! `set_opts( output = "excel" )` and `save_excel()` to generate 
+an Excel workbook, which consists of one or more spreadsheets.
 
 # surveytable 0.9.8
 
@@ -22,7 +24,6 @@ value of interest.
 
 * Ability to customize how the tables are printed.
   * `output` argument to `set_opts()`
-  * New function: `as_object()`
   * More details in `help("surveytable-options")`
 * Ability to customize rounding.
   * More details in `help("surveytable-options")`